+++ indicators/core/nuts_utils.py
@@ -8,7 +8,8 @@ Tools for dealing with NUTS regions.
 from functools import lru_cache
 from indicators.core.config import EU_COUNTRIES
 from indicators.core.config import NUTS_EDGE_CASES
-
+from itertools import groupby
+from operator import itemgetter
 from nuts_finder import NutsFinder as _NutsFinder
 
 
@@ -60,3 +61,49 @@ def iso_to_nuts(iso_code):
     elif iso_code == "GR":
         nuts_code = "EL"
     return nuts_code
+
+
+def make_reverse_lookup(data, key=itemgetter(1), prefix=""):
+    """Group an iterable of the form (`id`, `code_object`),
+    where a `code` exists somewhere in `code_object` (extracted
+    with `key`) such that the output is of the form:
+
+       {code: {id1, id2, id3}}
+
+    You might think of this as being {nuts_code: {article ids}}
+
+    Args:
+      data: Iterable of the form (`id`, `code_object`)
+      key: Function for extracting `code` from `code_object`
+      prefix: Prefix for the code, if required (intended to distinguish
+              NUTS from ISO codes) (Default value = "")
+
+    Returns:
+       Grouped object of the form {code: {id1, id2, id3}}
+    """
+    # Look nuts code --> article id
+    return {
+        f"{prefix}{code}": set(id for id, _ in group)
+        for code, group in groupby(sorted(data, key=key), key=key)
+        if code is not None  # Not interested in null codes
+    }
+
+
+@lru_cache()
+def get_geo_lookup(lat_lon_getter, iso2_to_id_getter):
+    # Forward lookup
+    nf = NutsFinder()
+    id_to_nuts_lookup = {
+        id: nf.find(lat=lat, lon=lon) for id, lat, lon in lat_lon_getter()
+    }
+    id_nuts = [  # splatten out the nuts IDs, ready for grouping
+        (id, info["NUTS_ID"])
+        for id, nuts_info in id_to_nuts_lookup.items()
+        for info in nuts_info
+    ]
+    id_iso2 = iso2_to_id_getter()
+    # Reverse lookups
+    nuts_to_id_lookup = make_reverse_lookup(id_nuts)
+    iso2_to_id_lookup = make_reverse_lookup(id_iso2, prefix="iso_")
+    # Combine lookups and return
+    return {**nuts_to_id_lookup, **iso2_to_id_lookup}
