+++ indicators/two/arxiv_topics.py
@@ -6,13 +6,11 @@ Topic modelling of the arXiv data.
 """
 
 from functools import lru_cache
-from itertools import groupby
-from operator import itemgetter
 
 from indicators.core.config import EU_COUNTRIES, ARXIV_CONFIG
 from indicators.core.db import get_mysql_engine
-from indicators.core.nlp_utils import fit_topics, vectorise_docs
-from indicators.core.nuts_utils import NutsFinder
+from indicators.core.nlp_utils import fit_topic_model
+from indicators.core.nuts_utils import get_geo_lookup
 from nesta.core.orms.arxiv_orm import Article as Art
 from nesta.core.orms.arxiv_orm import ArticleInstitute as Link
 from nesta.core.orms.grid_orm import Institute as Inst
@@ -42,33 +40,7 @@ def get_arxiv_eu_insts():
         return list(q.all())
 
 
-def make_reverse_lookup(data, key=itemgetter(1), prefix=""):
-    """Group an iterable of the form (`id`, `code_object`),
-    where a `code` exists somewhere in `code_object` (extracted
-    with `key`) such that the output is of the form:
-
-       {code: {id1, id2, id3}}
-
-    You might think of this as being {nuts_code: {article ids}}
-
-    Args:
-      data: Iterable of the form (`id`, `code_object`)
-      key: Function for extracting `code` from `code_object`
-      prefix: Prefix for the code, if required (intended to distinguish
-              NUTS from ISO codes) (Default value = "")
-
-    Returns:
-       Grouped object of the form {code: {id1, id2, id3}}
-    """
-    # Look nuts code --> article id
-    return {
-        f"{prefix}{code}": set(id for id, _ in group)
-        for code, group in groupby(sorted(data, key=key), key=key)
-        if code is not None  # Not interested in null codes
-    }
-
-
-def get_iso2_to_id_lookup():
+def get_iso2_to_id():
     """
     Fetch and curate a lookup table of ISO2 code to
     the set of article ids for that ISO2 code.
@@ -77,34 +49,7 @@ def get_iso2_to_id_lookup():
     with db_session(engine) as session:
         q = session.query(Link.article_id, Inst.country_code)
         q = q.join(Link, Link.institute_id == Inst.id, isouter=True)
-        iso2_to_id_lookup = make_reverse_lookup(q.all(), prefix="iso_")
-    return iso2_to_id_lookup
-
-
-@lru_cache()
-def get_arxiv_geo_lookup():
-    """Generate a lookup table of geography codes (NUTS/ISO2)
-    to arXiv article IDs.
-
-    Returns:
-        geo_to_id_lookup (dict): Lookup of geography codes to arXiv article IDs.
-    """
-    # Forward lookup
-    nf = NutsFinder()
-    id_to_nuts_lookup = {
-        id: nf.find(lat=lat, lon=lon) for id, lat, lon in get_arxiv_eu_insts()
-    }
-    id_nuts = [  # splatten out the nuts IDs, ready for grouping
-        (id, info["NUTS_ID"])
-        for id, nuts_info in id_to_nuts_lookup.items()
-        for info in nuts_info
-    ]
-    # Reverse lookups
-    nuts_to_id_lookup = make_reverse_lookup(id_nuts)
-    iso2_to_id_lookup = get_iso2_to_id_lookup()
-    # Combine lookups and return
-    nuts_to_id_lookup.update(**iso2_to_id_lookup)
-    return nuts_to_id_lookup
+        return list(q.all())
 
 
 @lru_cache()
@@ -146,7 +91,7 @@ def get_arxiv_articles(from_date="2015-01-01", geo_split=False):
     """
     articles = _get_arxiv_articles(from_date=from_date)
     if geo_split:
-        nuts_to_id_lookup = get_arxiv_geo_lookup()
+        nuts_to_id_lookup = get_geo_lookup(get_arxiv_eu_insts, get_iso2_to_id)
         for geo_code, ids in nuts_to_id_lookup.items():
             indexes = list(article["id"] in ids for article in articles)
             yield indexes, geo_code
@@ -154,25 +99,5 @@ def get_arxiv_articles(from_date="2015-01-01", geo_split=False):
         yield articles
 
 
-def fit_arxiv_topics():
-    """Fit arXiv topics based on hyperparameters specified in 'arxiv.yaml'.
-
-    Returns:
-        articles, topic_model: List of articles, and a trained topic model
-    """
-    articles = next(get_arxiv_articles())  # only one value (a list), so use next
-    texts = [art["text"] for art in articles]
-    titles = [art["title"] for art in articles]
-    # Prepare the data and fit the model
-    doc_vectors, feature_names = vectorise_docs(texts)
-    topic_model = fit_topics(
-        titles=titles,
-        doc_vectors=doc_vectors,
-        feature_names=feature_names,
-        **ARXIV_CONFIG,
-    )
-    return articles, topic_model
-
-
 if __name__ == "__main__":
-    fit_arxiv_topics()
+    fit_topic_model(ARXIV_CONFIG, get_arxiv_articles)
